Prio: flesh out "What is Prio"
Adds a paragraph or two offering background on Prio, links to the
Stanford website and paper and a very high level description of how the
system works.

Also replaces usage of "facilitation" or "facilitation server" with
"data share processor" to be more consistent with the terminology we
want to adopt in design docs and source code.